Moved websocket class to separated file
